Move fetchUserInfo logic to the providers pkg
We used to control how the ID token was parsed into claims in the broker
itself. This would become progressively unmaintainable as each provider
could understand the claims differently. Now, instead of having to
enforce the providers to use the email as username for local
authentication, they control what will be parsed from the ID Token and
return a UserInfo to broker, which then only fills the empty values with
defaults (if any).