Store token unencrypted
Before this commit, we encrypted the token with the local password
before writing it to the filesystem. We decrypted the token whenever the
user logs in and try to refresh it. That required us to always ask for
the local password on login, even if the user authenticated via OIDC.

We discussed various approaches to solve this issue. In the end we
decided to go with the simplest approach, i.e. to store the token
unencrypted. It is still protected from unprivileged users via
filesystem permissions.

To still be able to verify the local password, we now store a hash of
the local password in a file.

For a seamless transition, we still support logging in via a local password
when there is no password file but an old encrypted token file. This
should make the change completely transparent to users.

This commit also uses the term "data directory" for what was previously
known as the "cache directory", because we're storing the tokens and
password file in there, which should not be considered "cache" (because
if the user deletes those, they won't be able to log in with the local
password anymore).