Disable the thelper linter
It's not useful to always call t.Helper() in all functions which receive
a `t *testing.T` argument. For example, when using a `setup()` function
in test cases, we don't want that setup function to be skipped when
printing the file and line information in a log message, because that
function is only called exactly once by exactly one test case, so no
information is lost.

Disabling the linter for each `setup()` function would require adding a
lot of `//nolint` directives which clutter our code. Lets completely
disable the thelper linter instead.