msentraid: Support registering device upon login using libhimmelblau
Device registration enables Entra administrators to monitor which users
have logged in on which devices. It’s also a prerequisite for Single
Sign-On (SSO), which we plan to support in the future.

This change adds a register_device setting to the authd configuration,
which controls automatic device registration upon user login.

If the device is disabled in Microsoft Entra, user login is denied.

Closes https://github.com/ubuntu/authd/issues/697