Fix invalid msgraph host
Since commit 5b6a01d748e809499562b1c585ac4e030bb99eaf, we store the full
URL to the msgraph host in the provider metadata. However, the metadata
on disk is not updated, so it can still contain only the domain name of
the msgraph host, resulting in a failure to get the groups via msgraph:

    Could not fetch user info: could not get user info: failed to get user groups: failed to get user groups: Get "graph.microsoft.com/me/transitiveMemberOf/graph.group": unsupported protocol scheme "". Using cached user info.

This commit fixes that by supporting both the old and new formats of the
msgraph host field in the provider metadata.

Closes #858