Remove support for old encrypted token
We don’t store the token encrypted anymore since commit
2a71efc880792807aadc9085fa4639731f154ddd which was merged on
Nov 7, 2024 and released in v0.2.0 of the broker on Jan 30, 2025.
That should be long enough ago that we don’t need to support it
anymore, all active users should have had re-authenticate via device
auth by now, which will have created a new, unencrypted token.

This simplifies things for the device registration PR, where we need to
parse the token in GetAuthenticationModes to decide if local password
authentication should be offered or not (it should not in case that
device registration is enabled in the broker config and the stored token
is an old token which can’t be used for device registration).