e2e-tests: Provide script for provisioning the VM
... and a e2e-tests/vm/ssh.sh script to connect to the VM via SSH.

The provisioning script does all the steps necessary before running the
tests, including installation and configuration of authd and the brokers
and creation of snapshots.

We now use public key authentication when connecting to the VM via SSH,
so the provisioning script requires an SSH public key as argument.

Using public key authentication allows us to avoid using sshpass and is
more secure, because it avoids that other libvirt VMs using the default
network can access the e2e-runner.